(PC-13051)[PRO] refactor: use improvements
