+++ pro/src/components/pages/Offerers/Offerer/VenueV1/VenueEdition/ImageVenueUploaderSection/ImageVenueUploaderSection.module.scss
@@ -15,9 +15,4 @@
     justify-content: space-evenly;
     margin-left: rem(32px);
   }
-
-  &-icon {
-    margin-right: rem(8px);
-    width: 20px;
-  }
 }
