+++ pro/src/components/pages/Offerers/Offerer/VenueV1/VenueEdition/ImageVenueUploaderSection/VenueImageUploaderModal.tsx
@@ -1,6 +1,7 @@
 import React, { useCallback, useState } from 'react'
 import { CroppedRect } from 'react-avatar-editor'
 
+import { getDataURLFromImageURL } from 'api/utils/api'
 import useNotification from 'components/hooks/useNotification'
 import { imageConstraints } from 'new_components/ConstraintCheck/imageConstraints'
 import DialogBox from 'new_components/DialogBox'
@@ -11,7 +12,6 @@ import { VenueImageEdit } from '../VenueImageEdit/VenueImageEdit'
 import { VenueImagePreview } from '../VenueImagePreview/VenueImagePreview'
 
 import { IMAGE_TYPES, MAX_IMAGE_SIZE, MIN_IMAGE_WIDTH } from './constants'
-import { getDataURLFromImageURL } from './utils'
 
 interface IVenueImageUploaderModalProps {
   venueId: string
@@ -72,22 +72,21 @@ export const VenueImageUploaderModal = ({
   }, [setImage])
 
   const onUpload = useCallback(async () => {
-    try {
-      if (typeof croppingRect === 'undefined') return
-      if (typeof image === 'undefined') return
+    if (typeof croppingRect === undefined) return
+    if (typeof image === undefined) return
 
+    try {
+      setIsUploading(true)
       // the request needs the dataURL of the image,
       // so we need to retrieve it if we only have the URL
       const imageDataURL =
         typeof image === 'string' ? await getDataURLFromImageURL(image) : image
-
-      setIsUploading(true)
       const { bannerUrl } = await postImageToVenue({
         venueId,
         banner: imageDataURL,
-        xCropPercent: croppingRect.x,
-        yCropPercent: croppingRect.y,
-        heightCropPercent: croppingRect.height,
+        xCropPercent: croppingRect?.x,
+        yCropPercent: croppingRect?.y,
+        heightCropPercent: croppingRect?.height,
         imageCredit: credit,
       })
       onImageUpload({ bannerUrl, credit })
@@ -96,7 +95,7 @@ export const VenueImageUploaderModal = ({
       notification.success('Vos modifications ont bien été prises en compte')
     } catch {
       notification.error(
-        'Une erreur est survenue. Merci de réessayer plus tard'
+        'Une erreur est survenue lors de la sauvegarde de vos modifications.\n Merci de réessayer plus tard'
       )
       setIsUploading(false)
     }
